Further improvements and fixes after merge

+++ include/XLink/XLink.h
@@ -291,7 +291,7 @@ XLinkError_t XLinkWriteData(streamId_t const streamId, const uint8_t* buffer, in
  * @param[in] buffer2Size – size of the data to be transmitted
  * @return Status code of the operation: X_LINK_SUCCESS (0) for success
  */
-XLinkError_t XLinkWriteData2(streamId_t streamId, const uint8_t* buffer1, size_t buffer1Size, const uint8_t* buffer2, size_t buffer2Size);
+XLinkError_t XLinkWriteData2(streamId_t streamId, const uint8_t* buffer1, int buffer1Size, const uint8_t* buffer2, int buffer2Size);
 
 /**
  * @brief Sends a package to initiate the writing of data to a remote stream
